Alternate implementation for closing stream race
This approach adds extra states for CLOSING to the close state machine that represent
when a endStream close from has been generated but not yet sent.   The state machine has
been modified in a way so that if we miss he closing states, it should still function as before.

If a closing state is entered, then a count of closing streams in maintained and used in the
calculation for maxRemoteStreams

Signed-off-by: Greg Wilkins <gregw@webtide.com>